requirements.txt
numpy~=1.21.4
scipy~=1.7.3
torch~=1.12.1
pandas~=1.3.5
matplotlib~=3.5.1
scikit-learn~=1.0.1
Pillow~=8.4.0
PyYAML~=6.0
torchvision~=0.13.1
pymia~=0.3.2
tensorboard~=2.9.1